Fix the about page
The address listed on our contact page is now current.

The Team page was listing almost every dream director on the DC page.
That's because when I added a label for "Dream Directors", I wisely
chose to call it `ddc`, which of course contains the string `dc`. And
guess what? The team listings use string matchers, not string equality
checks, to determine who should be there.

I fixed this by renaming the team `dd`, but that's hacky and it's
getting to be time for this whole site to be rebuilt, with a real CMS.